refactor: introduce backend app factory with injection seams (#106)
Extract createApp({ auth, stripe, mailer, prisma }) into backend/app.js so
the whole HTTP surface is testable without a port or real external
services; index.js shrinks to env validation, real wiring, and listen.

Auth is a grouped seam ({ middleware, requireAuth, getAuth }) because
Clerk's getAuth only reads state its own middleware sets — a fake
middleware alone can't impersonate. Prisma joins the factory params so
the DB-backed suite can point the app at the _test database.

Adds test fakes (auth impersonation, fake Stripe, capture-mailer) and a
first supertest smoke test on GET /products under npm run test:db.
