+++ backend/emails/index.ts
@@ -1,4 +1,5 @@
 export { sendEmail } from "./sendEmail.ts";
+export type { Mailer, SendEmailOptions } from "./sendEmail.ts";
 export { renderOrderConfirmation } from "./renderOrderConfirmation.ts";
 export { renderAdminNewOrder } from "./renderAdminNewOrder.ts";
 export { renderOrderShipped } from "./renderOrderShipped.ts";
