+++ backend/package.json
@@ -18,9 +18,9 @@
   "type": "module",
   "dependencies": {
     "@clerk/express": "^2.0.7",
-    "@sznyt/shared": "*",
     "@prisma/adapter-pg": "^7.4.2",
     "@prisma/client": "^7.5.0",
+    "@sznyt/shared": "*",
     "cors": "^2.8.6",
     "dotenv": "^17.3.1",
     "express": "^5.2.1",
@@ -32,7 +32,9 @@
   },
   "devDependencies": {
     "@types/nodemailer": "^7.0.11",
+    "@types/supertest": "^7.2.0",
     "nodemon": "^3.1.14",
+    "supertest": "^7.2.2",
     "vitest": "^4.1.10"
   }
 }
